+++ python/cudf/groupby/groupby.py
@@ -16,6 +16,40 @@ from libgdf_cffi import ffi, libgdf
 from librmm_cffi import librmm as rmm
 
 
+class SeriesGroupBy(object):
+    """Wraps DataFrameGroupby with special attr methods
+    """
+    def __init__(self, source_series, group_series):
+        self.source_series = source_series
+        self.group_series = group_series
+
+    def __getattr__(self, attr):
+        df = DataFrame()
+        df['x'] = self.source_series
+        df['y'] = self.group_series
+        groupby = df.groupby('y')
+        result_df = getattr(groupby, attr)()
+
+        def get_result():
+            result_series = result_df['x']
+            result_series.name = None
+            idx = result_df.index
+            idx.name = None
+            result_series.set_index(idx)
+            return result_series
+        return get_result
+
+    def agg(self, agg_types):
+        df = DataFrame()
+        df['x'] = self.source_series
+        df['y'] = self.group_series
+        groupby = df.groupby('y').agg(agg_types)
+        idx = groupby.index
+        idx.name = None
+        groupby.set_index(idx)
+        return groupby
+
+
 class Groupby(object):
     """Groupby object returned by cudf.DataFrame.groupby().
     """
@@ -108,6 +142,9 @@ class Groupby(object):
             if agg_type == "count":
                 out_col_agg_series = Series(
                     Buffer(rmm.device_array(col_agg.size, dtype=np.int64)))
+            elif agg_type == "mean":
+                out_col_agg_series = Series(
+                    Buffer(rmm.device_array(col_agg.size, dtype=np.float64)))
             else:
                 out_col_agg_series = Series(Buffer(rmm.device_array(
                     col_agg.size, dtype=self._df[val_col]._column.data.dtype)))
@@ -158,7 +195,7 @@ class Groupby(object):
 
         return result
 
-    def _apply_basic_agg(self, agg_type):
+    def _apply_basic_agg(self, agg_type, sort_results=False):
         """
         Parameters
         ----------
@@ -178,7 +215,7 @@ class Groupby(object):
 
         result = self._apply_agg(
             agg_type, result, add_col_values, ctx, val_columns,
-            val_columns_out, sort_result=False)
+            val_columns_out, sort_result=sort_results)
 
         # If a Groupby has one index column and one value column
         # and as_index is set, return a Series instead of a df
@@ -212,20 +249,20 @@ class Groupby(object):
             return self[key]
         raise AttributeError("'Groupby' object has no attribute %r" % key)
 
-    def min(self):
-        return self._apply_basic_agg("min")
+    def min(self, sort=True):
+        return self._apply_basic_agg("min", sort)
 
-    def max(self):
-        return self._apply_basic_agg("max")
+    def max(self, sort=True):
+        return self._apply_basic_agg("max", sort)
 
-    def count(self):
-        return self._apply_basic_agg("count")
+    def count(self, sort=True):
+        return self._apply_basic_agg("count", sort)
 
-    def sum(self):
-        return self._apply_basic_agg("sum")
+    def sum(self, sort=True):
+        return self._apply_basic_agg("sum", sort)
 
-    def mean(self):
-        return self._apply_basic_agg("mean")
+    def mean(self, sort=True):
+        return self._apply_basic_agg("mean", sort)
 
     def agg(self, args):
         """ Invoke aggregation functions on the groups.
@@ -266,8 +303,6 @@ class Groupby(object):
         use_prefix = 1 < len(self._val_columns) or 1 < len(args)
         if not isinstance(args, str) and isinstance(
                 args, collections.abc.Sequence):
-            if (len(args) == 1 and len(self._val_columns) == 1):
-                sort_result = False
             for agg_type in args:
                 val_columns_out = [agg_type + '_' +
                                    val for val in self._val_columns]